pyproject.toml
[project]
name = "dlthub-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[bigquery,duckdb,filesystem,gs,postgres,sftp,sql-database,workspace]>=1.21.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.19.0",
    "yfinance>=1.1.0",
    "altair<5",
    "google-cloud-bigquery-storage>=2.36.2",
]
